+++ src/main/java/com/dnd/moyeolak/domain/participant/service/ParticipantService.java
@@ -1,13 +1,15 @@
 package com.dnd.moyeolak.domain.participant.service;
 
-import com.dnd.moyeolak.domain.meeting.entity.Meeting;
+import com.dnd.moyeolak.domain.participant.dto.CreateParticipantResponse;
+import com.dnd.moyeolak.domain.participant.dto.CreateParticipantWithLocationRequest;
+import com.dnd.moyeolak.domain.participant.dto.CreateParticipantWithScheduleRequest;
 import com.dnd.moyeolak.domain.participant.entity.Participant;
 
 public interface ParticipantService {
 
+    CreateParticipantResponse createWithSchedule(String meetingId, CreateParticipantWithScheduleRequest request);
+
+    CreateParticipantResponse createWithLocation(String meetingId, CreateParticipantWithLocationRequest request);
+
     void save(Participant participant);
-
-    Participant create(Meeting meeting, String name, String localStorageKey);
-
-    void validateLocalStorageKeyUnique(Meeting meeting, String localStorageKey);
 }
